Add RoboState::globalDriveTime, a one-time-use double for modifying drive time.
This needs to be used when doing the calculation for how far the robot should
drive as calculated using vision. Since the drive will be executed after the
robot has already rotated, a second vision calculation may **not** be done. It
must instead be passed in through some outside method, hence this new variable.

As a safety measure, it outputs a log message explicitly warning the user that
the drive time for that state has been overwritten. For now I can't think of any other way to pass that time value in.