Fix "armPos = 0" by changing to "armPos = DOWN"
This means that drivers MUST tell the arm to go up before teleop starts if they
want to use the shooter macro, even if it already is up. However, I am
confident that they will at some point tell the arm to go *down* at some point
fisrt, and thus this will not be an issue.